third_party/blink/web_tests/platform/linux-chrome/external/wpt/resize-observer/svg-expected.txt

This is a testharness.js-based test.
[FAIL] test15: observe svg:text content and border box
  assert_equals: expected 360 but got 138.3125
Harness: the test ran to completion.
